+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Dashboard revamped, shines in a new style and new features !2227 !2245 !2248 !2252 !2254 !2266 @sefn
 - In the footer of the website, "For developers" is added under "Participate" to create an easier way to IT. !2241 @NINI1988
 - On the login page, "Stay logged in for one day" is saved in local storage for the next logins. !2249 @NINI1988
+- Food-share-points are listed in the topbar search #1251 !2260 @alex.simm 
 
 ## Changes
 - Checks in the id card creation function if only one foodsaver is selected and then reduces the